rename: rename busRental to rental
